Fixes #157

+++ src/main/java/nl/sidnlabs/entrada/metric/Metric.java
@@ -19,7 +19,7 @@ public interface Metric {
 
   String getName();
 
-  int getValue();
+  double getValue();
 
   void setCached();
